1.benchmark.h:add 'load_trace' function to replace previous load_binary/test_data(); 2.run.sh:add compile command and remove keys_file_type & table_size; 3.microbench.cpp:modify key/value type to uint32_t to adapt lisa datasets;

+++ src/benchmark/microbench.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char **argv)
 {
-    Benchmark<uint64_t, uint64_t> bench;
+    Benchmark<uint32_t, uint32_t> bench;
     bench.parse_args(argc, argv);
     bench.run_benchmark();
 }
